+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.2.0
 
 - Auto-update shall only run once in every 24 hours with help of the file `~/.libautoupdate`
+- add argument `force` to `autoupdate()` in order to force an auto-update check
+- add CI to test `autoupdate()`
 
 ## v0.1.0